refactor: change folder structure of tests
