+++ .secrets.baseline
@@ -4,7 +4,7 @@
     "files": "package-lock.json|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2021-01-11T08:57:41Z",
+  "generated_at": "2021-04-09T07:48:36Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -55,7 +55,7 @@
     }
   ],
   "results": {
-    "backend/tests/test-env.ts": [
+    "backend/src/test-utils/test-env.ts": [
       {
         "hashed_secret": "c237a19676ad55d8904be354990dd54f92f9572c",
         "is_verified": false,
@@ -66,7 +66,6 @@
     "frontend/.env-example": [
       {
         "hashed_secret": "e5e9fa1ba31ecd1ae84f75caaa474f3a663f05f4",
-        "is_secret": false,
         "is_verified": false,
         "line_number": 5,
         "type": "Basic Auth Credentials"
