+++ backend/tsconfig.json
@@ -57,7 +57,7 @@
       "@sms/*": ["src/sms/*"],
       "@email/*": ["src/email/*"],
       "@telegram/*": ["src/telegram/*"],
-      "@tests/*": ["tests/*"],
+      "@test-utils/*": ["src/test-utils/*"],
     },                           /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
     // "typeRoots": [],                       /* List of folders to include type definitions from. */
